move vtv model to own top-level module
