+++ vtv-model/actions/deleteNode.js
