+++ vtv-model/actions/deleteValue.js
