+++ vtv-model/actions/setHidden.js
