+++ vtv-model/actions/setView.js
