+++ vtv-model/actions/showAll.js
