+++ vtv-model/analyze.js
