+++ vtv-model/util/clearStateProperty.js
