+++ vtv-model/util/draftState.js
