+++ vtv-model/util/getDraftStateUpdate.js
