+++ vtv-model/util/getDraftUpdate.js
