+++ vtv-model/util/getNewKey.js
