+++ vtv-model/util/getState.js
