+++ vtv-model/util/removeTemporaryState.js
