+++ vtv/index.js
@@ -1,13 +1,5 @@
 import View from './View'
 export { default as Textarea } from './generic/Textarea'
-export {
-  getPaths,
-  splitPath,
-  joinPath,
-  parseCommand,
-  removeTemporaryState,
-  updateTree,
-} from './model'
 export { getTheme } from './themes'
 
 export default View
